# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Flaque API
  version: 0.2.0
  summary: API for the Flaque self-hosted audio platform
  description: |
    This specification documents the HTTP API exposed by the Flaque backend.

    - Base URL defaults to `http://localhost:4000`
    - Most endpoints require the `flaque_session` cookie obtained from `POST /api/auth/login`
    - Admin-only endpoints are explicitly marked in endpoint descriptions
servers:
  - url: http://localhost:4000
    description: Local development server
security:
  - cookieAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and create a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful, session cookie is set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthForgotPasswordRequest"
      responses:
        "200":
          description: Request accepted regardless of account existence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset account password using a reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthResetPasswordRequest"
      responses:
        "200":
          description: Password reset applied and all existing sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and clear current session
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get authenticated user
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/sessions:
    get:
      tags: [Auth]
      summary: List active sessions for the current user
      responses:
        "200":
          description: Active sessions ordered by most recent activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionsEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Revoke a specific session for the current user
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/auth/logout-others:
    post:
      tags: [Auth]
      summary: Revoke all sessions except the current one
      responses:
        "200":
          description: Number of revoked sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutOtherSessionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flaque_session
  parameters:
...
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NoContent:
      description: Request completed with no response body
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    UserRole:
      type: string
      enum: [admin, user]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required: [id, username, role]
    AuthUserEnvelope:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [user]
    LoginRequest:
      type: object
      properties:
        login:
          type: string
          description: Username or email
        username:
          type: string
          description: Deprecated alias for login
          deprecated: true
        password:
          type: string
        deviceLabel:
          type: string
          description: Optional human-readable client/device label
      required: [password]
      anyOf:
        - required: [login]
        - required: [username]
    AuthForgotPasswordRequest:
      type: object
      properties:
        login:
          type: string
          description: Username or email
      required: [login]
    AuthResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
      required: [token, newPassword]
    GenericOkResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
      required: [ok]
    AuthSession:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        createdAt:
          type: integer
        expiresAt:
          type: integer
        lastSeenAt:
          type: integer
        userAgent:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        label:
          type: ["string", "null"]
        current:
          type: boolean
      required: [id, userId, createdAt, expiresAt, lastSeenAt, userAgent, ipAddress, label, current]
    AuthSessionsEnvelope:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/AuthSession"
      required: [sessions]
    LogoutOtherSessionsResponse:
      type: object
      properties:
        revoked:
          type: integer
          minimum: 0
      required: [revoked]
...